Move tests to unit_tests directory
