+++ external-crates/move/crates/move-vm-runtime/src/unit_tests/mod.rs
@@ -11,6 +11,7 @@ mod exec_func_effects_tests;
 mod function_arg_tests;
 mod instantiation_tests;
 mod invariant_violation_tests;
+mod jump_table_tests;
 mod leak_tests;
 mod loader_tests;
 mod nested_loop_tests;
